quantity management

+++ TssOrder/src/main/java/com/TssCommerce/TssOrder/Service/OrderService.java
@@ -10,7 +10,7 @@ import java.util.List;
 
 public interface OrderService {
     public List<ProductDao> getProducts();
-    public ProductOrder addOrder(Long userId, HashMap<Long,Double> productIdUnitPriceList);
+    public ProductOrder addOrder(Long userId, HashMap<Long,Integer> productIdUnitPriceList);
     public ProductOrder getOrderById(Long id);
     public Double setTotalPrice(List<ProductDao> productDaoList);
     public Discount setDiscount(Double totalPrice);
